backend/requirements/dev.txt
-r base.txt

pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
django-debug-toolbar>=4.3,<5.0
ipython>=8.0,<9.0
